% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rmse.R
\name{cal_rmse}
\alias{cal_rmse}
\title{RMSE calculator function}
\usage{
cal_rmse(predicted, actual)
}
\arguments{
\item{predicted}{predicted a numeric vector representing the values}

\item{actual}{actual the numeric vector representing the actual values from data}
}
\value{
an int value indicating the RMSE
}
\description{
RMSE calculator function
}
\examples{
x <- seq(-10, 10, by = .1)
y <- dnorm(x, mean = 2.5, sd = 0.5)
cal_rmse(x,y)
}
