% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter plot function}
\usage{
scatter_plot(
  dataframe,
  x_var,
  y_var,
  title,
  x_label,
  y_label,
  axis_text_size = 20,
  axis_title_size = 20,
  plot_title_size = 25,
  legend = TRUE
)
}
\arguments{
\item{dataframe}{A data frame or data frame extension (e.g. a tibble).}

\item{x_var}{Unquoted column name of column containing the x values}

\item{y_var}{Unquoted column name of column containing the y values}

\item{title}{String containing the title of the plot}

\item{x_label}{String containing the name of the x axis}

\item{y_label}{String containing the name of the y axis}

\item{axis_text_size}{Integer to indicate the axis values size, default 20}

\item{axis_title_size}{Integer indicating the axis labels size, default 20}

\item{plot_title_size}{Integer indicating the title text size, default 25}

\item{legend}{Boolean to add/remove legend, default TRUE}
}
\value{
A ggplot2 scatter plot instance
}
\description{
Given a dataframe, two columns and other tunning parameters,
creates a scatter plot
}
\examples{
scatter_plot(
    df, Height, Width,
    "Heigth to Width comparison", "Heigth", "Width"
)
}
