% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split data into 3 data frames}
\usage{
split_data(dataset, train_perc = 0.6, vs_prec = 0.2, test_perc = 0.2)
}
\arguments{
\item{dataset}{a data set or data frame}

\item{train_perc}{percentage of train data, default value is 0.6}

\item{vs_prec}{percentage of variable selection data, default value is 0.2}

\item{test_perc}{percentage of test data, default value is 0.2}
}
\value{
a list of 3 data frame: train data sets, test data sets, variable selection data sets
}
\description{
Split data into 3 data frames
}
\examples{
x=data.frame(replicate(10,sample(0:1,1000,rep=TRUE)))
out = split_data(x)


}
